pull compilation code out of compile task
