+++ dbt/task/test.py
@@ -1,0 +1,8 @@
+
+class TestTask:
+    def __init__(self, args, project):
+        self.args = args
+        self.project = project
+
+    def run(self):
+        print "Running tests!"
